Refactor: Rename AndroidLib package name from "com.sriky.jokeproviderandroidlib" to "com.sriky.jokedisplay"
