+++ jokeDisplayAndroidLib/src/main/AndroidManifest.xml
@@ -13,10 +13,10 @@
   ~ See the License for the specific language governing permissions and limitations under the License.
 -->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.sriky.jokeproviderandroidlib">
+    package="com.sriky.jokedisplay">
 
     <application>
-        <activity android:name=".JokeActivity"></activity>
+        <activity android:name="com.sriky.jokedisplay.JokeActivity"></activity>
     </application>
 
 </manifest>
