+++ jokeDisplayAndroidLib/src/main/res/values/strings.xml
