+++ jokeDisplayAndroidLib/src/test/java/com/sriky/jokedisplay/ExampleUnitTest.java
@@ -13,29 +13,20 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-package com.sriky.jokeproviderandroidlib;
-
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+package com.sriky.jokedisplay;
 
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
- * Instrumented test, which will execute on an Android device.
+ * Example local unit test, which will execute on the development machine (host).
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
+public class ExampleUnitTest {
     @Test
-    public void useAppContext() throws Exception {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getTargetContext();
-
-        assertEquals("com.sriky.jokeproviderandroidlib.test", appContext.getPackageName());
+    public void addition_isCorrect() throws Exception {
+        assertEquals(4, 2 + 2);
     }
 }
